'Refactored by Sourcery'

+++ julia/modules/Anti-Spam_😈.py
@@ -61,46 +61,42 @@ async def is_register_admin(chat, user):
 
 @register(pattern="^/cleanbluetext ?(.*)")
 async def _(event):
-    if event.is_group:
-        if not await can_change_info(message=event):
-            return
-    else:
+    if (
+        event.is_group
+        and not await can_change_info(message=event)
+        or not event.is_group
+    ):
         return
-    args = event.pattern_match.group(1)
-    if args:
+    if args := event.pattern_match.group(1):
         val = args
         if val in ("off", "no"):
             sql.set_cleanbt(event.chat_id, False)
             reply = "Bluetext cleaning has been disabled for <b>{}</b>".format(
                 html.escape(event.chat.title)
             )
-            await event.reply(reply, parse_mode="html")
-
         elif val in ("yes", "on"):
             sql.set_cleanbt(event.chat_id, True)
             reply = "Bluetext cleaning has been enabled for <b>{}</b>".format(
                 html.escape(event.chat.title)
             )
-            await event.reply(reply, parse_mode="html")
-
         else:
             reply = "Invalid argument.Accepted values are 'yes', 'on', 'no', 'off'"
-            await event.reply(reply, parse_mode="html")
     else:
         clean_status = sql.is_enabled(event.chat_id)
         clean_status = "Enabled" if clean_status else "Disabled"
         reply = "Bluetext cleaning for <b>{}</b> : <b>{}</b>".format(
             event.chat.title, clean_status
         )
-        await event.reply(reply, parse_mode="html")
+    await event.reply(reply, parse_mode="html")
 
 
 @register(pattern="^/ignorecleanbluetext ?(.*)")
 async def _(event):
-    if event.is_group:
-        if not await can_change_info(message=event):
-            return
-    else:
+    if (
+        event.is_group
+        and not await can_change_info(message=event)
+        or not event.is_group
+    ):
         return
     args = event.pattern_match.group(1)
     chat = event.chat
@@ -122,10 +118,11 @@ async def _(event):
 
 @register(pattern="^/unignorecleanbluetext ?(.*)")
 async def _(event):
-    if event.is_group:
-        if not await can_change_info(message=event):
-            return
-    else:
+    if (
+        event.is_group
+        and not await can_change_info(message=event)
+        or not event.is_group
+    ):
         return
     args = event.pattern_match.group(1)
     chat = event.chat
@@ -148,12 +145,12 @@ async def _(event):
 @register(pattern="^/listcleanbluetext$")
 async def _(event):
 
-    if event.is_group:
-        if not await can_change_info(message=event):
-            return
-    else:
+    if (
+        event.is_group
+        and not await can_change_info(message=event)
+        or not event.is_group
+    ):
         return
-
     chat = event.chat
 
     global_ignored_list, local_ignore_list = sql.get_all_ignored(chat.id)
@@ -186,13 +183,11 @@ async def _(event):
     for ch in approved_userss:
         iid = ch['id']
         userss = ch['user']
-    if event.is_group:
-        if (await is_register_admin(event.input_chat, event.message.sender_id)):
-            return
-        if str(event.sender_id) in str(userss) and str(event.chat_id) in str(iid):
-            return
-        pass
-    else:
+    if not event.is_group:
+        return
+    if (await is_register_admin(event.input_chat, event.message.sender_id)):
+        return
+    if str(event.sender_id) in str(userss) and str(event.chat_id) in str(iid):
         return
     if str(event.sender_id) == "1246850012":
         return
@@ -233,16 +228,15 @@ async def profanity(event):
             "Please provide some input yes or no.\n\nCurrent setting is : **off**"
         )
         return
-    if input == "on":
-        if event.is_group:
-            chats = spammers.find({})
-            for c in chats:
-                if event.chat_id == c["id"]:
-                    await event.reply(
-                        "Profanity filter is already activated for this chat.")
-                    return
-            spammers.insert_one({"id": event.chat_id})
-            await event.reply("Profanity filter turned on for this chat.")
+    if input == "on" and event.is_group:
+        chats = spammers.find({})
+        for c in chats:
+            if event.chat_id == c["id"]:
+                await event.reply(
+                    "Profanity filter is already activated for this chat.")
+                return
+        spammers.insert_one({"id": event.chat_id})
+        await event.reply("Profanity filter turned on for this chat.")
     if input == "off":
         if event.is_group:  
             chats = spammers.find({})
@@ -254,7 +248,7 @@ async def profanity(event):
                     return
         await event.reply(
                     "Profanity filter isn't turned on for this chat.")
-    if not input == "on" and not input == "off":
+    if input not in ["on", "off"]:
         await event.reply("I only understand by on or off")
         return
 
@@ -281,16 +275,15 @@ async def profanity(event):
             "Please provide some input yes or no.\n\nCurrent setting is : **off**"
         )
         return
-    if input == "on":
-        if event.is_group:
-            chats = globalchat.find({})
-            for c in chats:
-                if event.chat_id == c["id"]:
-                    await event.reply(
-                        "Global mode is already activated for this chat.")
-                    return
-            globalchat.insert_one({"id": event.chat_id})
-            await event.reply("Global mode turned on for this chat.")
+    if input == "on" and event.is_group:
+        chats = globalchat.find({})
+        for c in chats:
+            if event.chat_id == c["id"]:
+                await event.reply(
+                    "Global mode is already activated for this chat.")
+                return
+        globalchat.insert_one({"id": event.chat_id})
+        await event.reply("Global mode turned on for this chat.")
     if input == "off":
         if event.is_group:  
             chats = globalchat.find({})
@@ -302,7 +295,7 @@ async def profanity(event):
                     return
         await event.reply(
                     "Global mode isn't turned on for this chat.")
-    if not input == "on" and not input == "off":
+    if input not in ["on", "off"]:
         await event.reply("I only understand by on or off")
         return
 
@@ -330,16 +323,15 @@ async def cleanservice(event):
             "Please provide some input yes or no.\n\nCurrent setting is : **off**"
         )
         return
-    if input in "on":
-        if event.is_group:
-            chats = cleanservices.find({})
-            for c in chats:
-                if event.chat_id == c["id"]:
-                    await event.reply(
-                        "Clean service message already enabled for this chat.")
-                    return
-            cleanservices.insert_one({"id": event.chat_id})
-            await event.reply("I will clean all service messages from now.")
+    if input in "on" and event.is_group:
+        chats = cleanservices.find({})
+        for c in chats:
+            if event.chat_id == c["id"]:
+                await event.reply(
+                    "Clean service message already enabled for this chat.")
+                return
+        cleanservices.insert_one({"id": event.chat_id})
+        await event.reply("I will clean all service messages from now.")
     if input in "off":
         if event.is_group:
             chats = cleanservices.find({})
@@ -351,8 +343,8 @@ async def cleanservice(event):
                     return
         await event.reply(
                     "Service message cleaning isn't turned on for this chat.")       
-    
-    if not input == "on" and not input == "off":
+
+    if input not in ["on", "off"]:
         await event.reply("I only understand by on or off")
         return
 
@@ -365,40 +357,41 @@ async def del_profanity(event):
     msg = str(event.text)
     sender = await event.get_sender()
     let = sender.username
-    if event.is_group:
-        if (await is_register_admin(event.input_chat, event.message.sender_id)):
-            return
-        pass
+    if event.is_group and (
+        await is_register_admin(event.input_chat, event.message.sender_id)
+    ):
+        return
     chats = spammers.find({})
     for c in chats:
-        if event.text:
-            if event.chat_id == c['id']:
-                if better_profanity.profanity.contains_profanity(msg):
-                    await event.delete()
-                    if sender.username is None:
-                        st = sender.first_name
-                        hh = sender.id
-                        final = f"[{st}](tg://user?id={hh}) **{msg}** is detected as a slang word and your message has been deleted"
-                    else:
-                        final = f'@{let} **{msg}** is detected as a slang word and your message has been deleted'
-                    dev = await event.respond(final)
-                    await asyncio.sleep(10)
-                    await dev.delete()
-        if event.photo:
-            if event.chat_id == c['id']:
-                await event.client.download_media(event.photo, "nudes.jpg")
-                if nude.is_nude('./nudes.jpg'):
-                    await event.delete()
-                    if sender.username is None:
-                        st = sender.first_name
-                        hh = sender.id
-                        final = f"[{st}](tg://user?id={hh}) your message has been deleted due to pornographic content"
-                    else:
-                        final = f'@{let} your message has been deleted due to pornographic content'
-                    dev = await event.respond(final)
-                    await asyncio.sleep(10)
-                    await dev.delete()
-                    os.remove("nudes.jpg")
+        if (
+            event.text
+            and event.chat_id == c['id']
+            and better_profanity.profanity.contains_profanity(msg)
+        ):
+            await event.delete()
+            if sender.username is None:
+                st = sender.first_name
+                hh = sender.id
+                final = f"[{st}](tg://user?id={hh}) **{msg}** is detected as a slang word and your message has been deleted"
+            else:
+                final = f'@{let} **{msg}** is detected as a slang word and your message has been deleted'
+            dev = await event.respond(final)
+            await asyncio.sleep(10)
+            await dev.delete()
+        if event.photo and event.chat_id == c['id']:
+            await event.client.download_media(event.photo, "nudes.jpg")
+            if nude.is_nude('./nudes.jpg'):
+                await event.delete()
+                if sender.username is None:
+                    st = sender.first_name
+                    hh = sender.id
+                    final = f"[{st}](tg://user?id={hh}) your message has been deleted due to pornographic content"
+                else:
+                    final = f'@{let} your message has been deleted due to pornographic content'
+                dev = await event.respond(final)
+                await asyncio.sleep(10)
+                await dev.delete()
+                os.remove("nudes.jpg")
 
 @tbot.on(events.NewMessage(pattern=None))
 async def del_profanity(event):
@@ -409,27 +402,26 @@ async def del_profanity(event):
     msg = str(event.text)
     sender = await event.get_sender()
     let = sender.username
-    if event.is_group:
-        if (await is_register_admin(event.input_chat, event.message.sender_id)):
-            return
-        pass
+    if event.is_group and (
+        await is_register_admin(event.input_chat, event.message.sender_id)
+    ):
+        return
     chats = globalchat.find({})
     for c in chats:
-        if event.text:
-            if event.chat_id == c['id']:
-                a = TextBlob(msg)
-                b = a.detect_language()
-                if not b == "en":
-                    await event.delete()
-                    if sender.username is None:
-                        st = sender.first_name
-                        hh = sender.id
-                        final = f"[{st}](tg://user?id={hh}) you should only speak in english here !"
-                    else:
-                        final = f'@{let} you should only speak in english here !'
-                    dev = await event.respond(final)
-                    await asyncio.sleep(10)
-                    await dev.delete()
+        if event.text and event.chat_id == c['id']:
+            a = TextBlob(msg)
+            b = a.detect_language()
+            if b != "en":
+                await event.delete()
+                if sender.username is None:
+                    st = sender.first_name
+                    hh = sender.id
+                    final = f"[{st}](tg://user?id={hh}) you should only speak in english here !"
+                else:
+                    final = f'@{let} you should only speak in english here !'
+                dev = await event.respond(final)
+                await asyncio.sleep(10)
+                await dev.delete()
 
 
 @tbot.on(events.ChatAction())
